Add LICENSE, README and update build.gradle file
Addition of an MIT license to the project. README was created to provide complete information about the project including the demo of how to call several OpenAI services without using the Spring AI project. Updated the 'org.springframework.boot' and 'io.netty:netty-resolver-dns-native-macos' versions in the build.gradle file for compatibility and security improvements.