Add ImageRequest creation method in OpenAIInterface
A new method has been added to the OpenAIInterface to create ImageRequests. This allows for easier creation and handling of ImageRequest objects in different services and tests. Refactored imports in OpenAIService to include all necessary JSON classes in one line. Added new tests in OpenAIServiceTest for creating ChatRequest and ImageRequest from defaults, checking that the created objects have correct values. Adjusted the call in OpenAIInterfaceTest to use newly created method in interface.

These changes improve the maintainability and simplify the object creation process.