Refactor OpenAIService and improve tests
This commit makes minor adjustments in the OpenAIService and its corresponding tests. Changes in the OpenAIServiceTest include removing redundant sorting of model names and changing the expected model in the assert statement. The AppConfig class is also refactored with a reduction in logging code for request headers. Plus, the getModelNames method in OpenAIService class now directly sorts the list of models.