Add new test, logger, and refactor method in OllamaService
In this commit, a new unit test 'conversation1' was added in OllamaServiceTest. Also, a logger was instantiated in 'OllamaService' to log responses from conversations. The method 'generate' in 'OllamaService' was refactored for cleaner and easier reading. Furthermore, some improvements were made to the JSON property naming and the JVM arguments for the test tasks were updated in the build script.